# app.py
import requests
from fastapi import FastAPI, Query
from xml.etree import ElementTree as ET
from scholarly import scholarly

app = FastAPI()

# -------- PubMed検索 --------
@app.get("/search_papers")
def search_papers(keyword: str = Query(..., description="検索するキーワード")):
    # PubMed検索
    url = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esearch.fcgi"
    params = {
        "db": "pubmed",
        "term": keyword,
        "retmode": "json",
        "retmax": 10
    }
    response = requests.get(url, params=params)
    data = response.json()

    pmids = data.get('esearchresult', {}).get('idlist', [])
    results = []

    # PubMedから詳細情報取得
    for pmid in pmids:
        fetch_url = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi"
        fetch_params = {
            "db": "pubmed",
            "id": pmid,
            "retmode": "xml"
        }
        fetch_response = requests.get(fetch_url, params=fetch_params)
        if fetch_response.status_code == 200:
            try:
                root = ET.fromstring(fetch_response.text)
                title = root.findtext(".//ArticleTitle") or "タイトル取得失敗"
                doi = root.findtext(".//ArticleId[@IdType='doi']")
            except Exception:
                title = "タイトル取得失敗"
                doi = "DOI not found"
        else:
            title = "タイトル取得失敗"
            doi = "DOI not found"

        # Crossrefからメタデータを取得するためにDOIを渡す
        metadata = get_crossref_metadata(doi) if doi != "DOI not found" else {}

        results.append({
            "pmid": pmid,
            "title": title,
            "doi": doi,
            "link": f"https://pubmed.ncbi.nlm.nih.gov/{pmid}/",
            "crossref_metadata": metadata
        })

    # Google Scholar から検索結果を取得
    google_scholar_papers = get_google_scholar_papers(keyword)

    # bioRxiv から検索結果を取得
    biorxiv_papers = get_biorxiv_papers(keyword)

    # 結果を統合
    all_papers = results + google_scholar_papers + biorxiv_papers

    return {"papers": all_papers}


# -------- Google Scholar検索 --------
def get_google_scholar_papers(keyword):
    search_query = scholarly.search_pubs(keyword)
    results = []
    for pub in search_query:
        results.append({
            "title": pub.get('bib', {}).get('title', 'No title'),
            "link": pub.get('url', 'No link')
        })
    return results


# -------- bioRxiv検索 --------
def get_biorxiv_papers(keyword):
    url = f"https://api.biorxiv.org/details/2022/03/01/{keyword}/json"
    response = requests.get(url)
    data = response.json()
    
    results = []
    for item in data['collection']:
        results.append({
            "title": item.get('title', 'No title'),
            "link": item.get('link', 'No link')
        })
    return results


# -------- Crossref検索 --------
def get_crossref_metadata(doi):
    if doi == "DOI not found":
        return {}

    url = f"https://api.crossref.org/works/{doi}"
    response = requests.get(url)
    if response.status_code != 200:
        return {"error": f"DOI {doi} not found"}

    data = response.json()["message"]
    return {
        "title": data.get("title", [""])[0],
        "authors": [f"{a.get('given', '')} {a.get('family', '')}" for a in data.get("author", [])],
        "journal": data.get("container-title", [""])[0],
        "published": data.get("issued", {}).get("date-parts", [[""]])[0],
        "abstract": data.get("abstract", "（抄録はありません）")
    }


@app.get("/get_metadata")
def get_metadata(doi: str = Query(..., description="DOI (Digital Object Identifier)")):
    return get_crossref_metadata(doi)