# openapi.yaml
openapi: 3.1.0
info:
  title: PubMed論文検索API
  version: 1.0.0
servers:
  - url: https://pubmed-gpt-action.onrender.com
paths:
  /search_papers:
    get:
      operationId: searchPapers
      summary: 複数のデータベース（PubMed、Google Scholar、bioRxiv）で論文を検索する
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
          description: 検索したいキーワード
      responses:
        "200":
          description: 検索結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  papers:
                    type: array
                    items:
                      type: object
                      properties:
                        pmid:
                          type: string
                        title:
                          type: string
                        link:
                          type: string
                        doi:
                          type: string
                        crossref_metadata:
                          type: object
                          properties:
                            title:
                              type: string
                            authors:
                              type: array
                              items:
                                type: string
                            journal:
                              type: string
                            published:
                              type: array
                              items:
                                type: string
                            abstract:
                              type: string
                        google_scholar_link:
                          type: string
                        biorxiv_link:
                          type: string

  /get_metadata:
    get:
      operationId: getMetadata
      summary: Crossref APIからDOIを基に論文メタデータを取得
      parameters:
        - name: doi
          in: query
          required: true
          schema:
            type: string
          description: DOI（例：10.1038/s41586-020-2649-2）
      responses:
        "200":
          description: Crossref論文情報
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  authors:
                    type: array
                    items:
                      type: string
                  journal:
                    type: string
                  published:
                    type: array
                    items:
                      type: string
                  abstract:
                    type: string